GRHB-42: * moved common folder out of helpers folder
