+++ frontend/src/components/uam/common/enums/enums.ts
