+++ frontend/src/components/uam/common/enums/user-table-accessor.enum.ts
