+++ frontend/src/components/uam/helpers/get-columns-helper/get-columns.helper.ts
@@ -1,7 +1,7 @@
 import { UsersGetAllItemResponseDto } from 'common/types/types';
 import { Column } from 'react-table';
 
-import { UserTableAccessor, UserTableHeader } from '../common/enums/enums';
+import { UserTableAccessor, UserTableHeader } from '../../common/enums/enums';
 
 const getColumns = (): Column<UsersGetAllItemResponseDto>[] => {
   return [
