+++ frontend/src/components/uam/helpers/get-rows-helper/get-rows.helper.ts
@@ -1,6 +1,6 @@
 import { UsersGetAllItemResponseDto } from 'common/types/types';
 
-import { UserTableAccessor } from '../common/enums/enums';
+import { UserTableAccessor } from '../../common/enums/enums';
 import { getTimeDifferenceWithCurrentTime } from '../helpers';
 
 const getRows = (
